(Partial) Implementation of Re-design of project properties.
After some discussion, it was decided that the way projects handle properties
should look like QSettings.  The learning curve would be smaller since the
interface would be more consistent with QSettings, which some developers would
be familiar with.

The read*() and write*() property functions do update the project internal
state.  However, the write*() has a bug with string lists in that the string
list entries are redundantly emitted to the project file.

Unfortunately reading this new property configuration from project files isn't
finished yet.  However, in the interest of getting something out there for
people to code against, I'm going ahead and committing what I have.  After
all, I want to reduce the number of people using the old interface, and thus
code that will have to be re-written to adjust to the new one.


git-svn-id: https://svn.osgeo.org/qgis/trunk/qgis@2282 c8812cc2-4d05-0410-92ff-de0c093fc19c